src/bot.py
import os
import logging
import asyncio
from datetime import datetime, timedelta
from telegram import Update
from telegram.ext import Application, CommandHandler, CallbackQueryHandler, ContextTypes
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    level=logging.INFO
)
logger = logging.getLogger(__name__)

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Send a message when the command /start is issued."""
    welcome_message = (
        "👋 Welcome to the Trading Bot!\n\n"
        "I'll help you with trading signals and market analysis.\n\n"
        "Available commands:\n"
        "/signal - Get latest trading signal\n"
        "/performance - View performance metrics\n"
        "/tracking - Get weekly tracking report\n"
        "/help - Show this help message"
    )
    await update.message.reply_text(welcome_message)

async def help_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Send a message when the command /help is issued."""
    help_text = (
        "🤖 *Trading Bot Commands*\n\n"
        "*/signal* - Get latest trading signal\n"
        "*/performance* - View performance metrics\n"
        "*/tracking* - Get weekly tracking report\n"
        "*/help* - Show this help message\n\n"
        "For more information, visit our website."
    )
    await update.message.reply_text(help_text, parse_mode='Markdown')

async def signal(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Handle the /signal command."""
    await update.message.reply_text("🔔 Getting latest trading signal...")
    # Add your signal generation logic here

async def performance(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Handle the /performance command."""
    await update.message.reply_text("📊 Getting performance metrics...")
    # Add your performance tracking logic here

async def tracking(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Handle the /tracking command."""
    await update.message.reply_text("📈 Getting weekly tracking report...")
    # Add your tracking report logic here

async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Handle button callbacks."""
    query = update.callback_query
    await query.answer()
    # Add your button handling logic here

def main():
    """Start the bot."""
    # Create the Application and pass it your bot's token
    application = Application.builder().token(os.getenv('TELEGRAM_BOT_TOKEN')).build()

    # Add command handlers
    application.add_handler(CommandHandler("start", start))
    application.add_handler(CommandHandler("help", help_command))
    application.add_handler(CommandHandler("signal", signal))
    application.add_handler(CommandHandler("performance", performance))
    application.add_handler(CommandHandler("tracking", tracking))
    application.add_handler(CallbackQueryHandler(button_callback))

    # Start the Bot
    logger.info("Bot started successfully!")
    application.run_polling()

if __name__ == '__main__':
    main() 