requirements.txt
python-telegram-bot>=20.0
pandas
numpy
pytz
python-dotenv
requests
schedule
tradingview_ta
tensorflow
scikit-learn
joblib
beautifulsoup4
apscheduler
TA-Lib
newsapi-python
ccxt
flask
MetaTrader5
yfinance